Converted file structure to jsx
